# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "AUG22pt-ep1-dissolution2",
    "description": "<b>Swagger Doc EP Noroff Backend API</b>."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "scheme": "bearer",
      "description": "Bearer token"
    }
  },
  "paths": {
...
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Guest, User or Admin, no restrictions. Can get all products",
        "responses": {
          "200": {
            "description": "statusCode: 200, data: { result: \"product found\", products }"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Admin can add new product: productName, description, quantity, price, discount, imgUrl, categoryId, brandId",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductsPost"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, result: { message: \"New product has been added to the database\" }"
          },
          "403": {
            "description": "statusCode: 403, result: { error: \"Privileges Admin rights [Required] !!\"}"
          },
          "500": {
            "description": "statusCode: 500, result: { error: \"Require categoryId or brandId, not in database create new Category or Brand\", input: categoryId or brandId }"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "description": "Admin can update/change product { productId, productName, description, quantity, price, discount, imgUrl, categoryId, brandId }",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductsPut"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, result: { message: \"Updated/Changed post\", updateObject: { true/false on updateded values } }"
          },
          "304": {
            "description": "statusCode: 304, result: { message: \"No Updated/Changed to post done\", updateObject }"
          },
          "403": {
            "description": "statusCode: 403, result: { error: \"Privileges Admin rights [Required] !!\"}"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "description": "Admin can sett product (softe deleted)",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductsDeleted"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, result: { message: \"Product has been set active(soft) in the database\" }"
          },
          "403": {
            "description": "statusCode: 403, result: { error: \"Privileges Admin rights [Required] !!\"}"
          },
          "500": {
            "description": "statusCode\": 500, result: { error: \"Error setting product active(soft undeleteing) product from the database... !!\"}"
          }
        }
      }
    },
    "/products/{productId}": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "User or Admin, no restrictions. Can get one product by Id",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, data: { result: \"product found\", products }"
          },
          "500": {
            "description": "statusCode\": 500, result: {error: \"Require productId, Not a integer value\", input: productId }"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProductsPost": {
      "type": "object",
      "properties": {
        "productName": {
          "type": "string",
          "example": "Test_Product_Swagger"
        },
        "description": {
          "type": "string",
          "example": "A prosject API"
        },
        "quantity": {
          "type": "number",
          "example": 1
        },
        "price": {
          "type": "number",
          "example": 100
        },
        "discount": {
          "type": "number",
          "example": 0
        },
        "imgUrl": {
          "type": "string",
          "example": "http://notareal.png"
        },
        "categoryId": {
          "type": "number",
          "example": 1
        },
        "brandId": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "productName",
        "description",
        "quantity",
        "price",
        "discount",
        "imgUrl",
        "categoryId",
        "brandId"
      ]
    },
    "ProductsPut": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "number",
          "example": 1
        },
        "productName": {
          "type": "string",
          "example": "iPhone 6s Plus 16Gb"
        },
        "description": {
          "type": "string",
          "example": "3D Touch. 12MP photos. 4K video."
        },
        "quantity": {
          "type": "number",
          "example": 10
        },
        "price": {
          "type": "number",
          "example": 749
        },
        "discount": {
          "type": "number",
          "example": 0
        },
        "imgUrl": {
          "type": "string",
          "example": "http://images.restapi.co.za/products/product-iphone.png"
        },
        "categoryId": {
          "type": "number",
          "example": 1
        },
        "brandId": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "productId",
        "productName",
        "description",
        "quantity",
        "price",
        "discount",
        "imgUrl",
        "categoryId",
        "brandId"
      ]
    },
    "ProductsDeleted": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "number",
          "example": 1
        },
        "isdeleted": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "productId",
        "isdeleted"
      ]
    },
...
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ]
}